docs(docs-infra): move the navigation related files
* separate directory
* dedicated file for redirections
