+++ adev/src/app/app.config.ts
@@ -31,7 +31,7 @@ import {AnalyticsService} from './core/services/analytics/analytics.service';
 import {ContentLoader} from './core/services/content-loader.service';
 import {CustomErrorHandler} from './core/services/errors-handling/error-handler';
 import {ExampleContentLoader} from './core/services/example-content-loader.service';
-import {routerProviders} from './router_providers';
+import {routerProviders} from './routing/router_providers';
 
 export const appConfig: ApplicationConfig = {
   providers: [
